PR-B5b: Lift witness specificity helpers into deterministic pack
Move the six deterministic specificity helpers out of
atlas_brain.autonomous.tasks._b2b_specificity into a new
extracted_quality_gate/witness_pack.py:

  - surface_specificity_context
  - merge_specificity_contexts
  - specificity_signal_terms
  - evaluate_specificity_support
  - specificity_audit_snapshot
  - campaign_proof_terms_from_audit

All six are pure: they operate on already-fetched anchor / witness
rows + candidate text. No DB, no clock, no settings, no Atlas types.
The atlas-side _b2b_specificity.py is now a thin re-export wrapper
(216 LOC, was 755) so existing imports
(`from atlas_brain.autonomous.tasks._b2b_specificity import ...`)
continue to resolve unchanged.

Pack contract entry point:

  evaluate_witness_specificity(input, *, policy) -> QualityReport

returns a standard QualityReport whose findings carry the
blocking_issues / warnings the underlying helpers produced.
QualityPolicy.thresholds knobs:

  - min_anchor_hits (default 1)
  - require_anchor_support (default True)
  - require_timing_or_numeric_when_available (default False)
  - include_competitor_terms (default True)
  - allow_company_names (default depends on `surface`)

Atlas wrapper changes:

- _b2b_specificity.py becomes a slim adapter layer: re-exports the
  six lifted helpers; keeps `campaign_policy_audit_snapshot` (the
  PR-B4b adapter, unchanged), `latest_specificity_audit`, and
  `specificity_quality_summary` (atlas-side metadata helpers that
  read pre-computed audit dicts off row metadata -- not deterministic
  validators, so they stay).
- All six external import sites
  (`atlas_brain/services/campaign_quality.py`,
  `atlas_brain/autonomous/tasks/b2b_blog_post_generation.py`,
  `atlas_brain/autonomous/tasks/b2b_campaign_generation.py`)
  continue to work without changes via the re-exports.

Tests:

- tests/test_extracted_quality_gate_witness_pack.py (30 tests):
  surface_specificity_context redaction, merge dedup,
  signal_terms group extraction, evaluate_specificity_support
  pass / block / warn paths, audit snapshot fields, proof-term
  ordering with channel + blocking-issue gates, the pack contract
  entry point (frozen result, policy override, surface-driven
  defaults), and atlas re-export identity.
- 286 tests pass across the affected suites
  (test_b2b_blog_quality_gate / test_extracted_quality_gate_* /
   test_b2b_campaign_generation / test_campaign_specificity_backfill /
   test_b2b_witnesses).

PR-B5 split:
  PR-B5b (this PR) is the foundation slice -- pure logic, biggest
  test ROI, lowest risk. PR-B5a (evidence coverage gate, async DB)
  and PR-B5c (source-quality pack, async DB + settings) follow.

Coordination:
- Owner claude-2026-05-03-b (alias A); claimed PR-B5 in queue.md
  with the three-way split note.
